+++ src/components/unread-dot/style.scss
@@ -1,0 +1,27 @@
+.notifications {
+
+	.unread-dot {
+		position: absolute;
+		width: 6px;
+		height: 6px;
+		top: 12px;
+		left: 20px;
+		border-radius: 50%;
+		font-size: 0;
+
+		/* the dot color */
+		background: $color-red;
+
+		/* the dot color border */
+		border: 1px solid $color-black-800;
+
+		/* initially it is hidden */
+		opacity: 0%;
+
+		transition: opacity 0.25s;
+
+		&.has-unread {
+			opacity: 100%;
+		}
+	}
+}
