+++ src/constants.ts
@@ -1,0 +1,19 @@
+/**
+ * WP Notification Feature namespace.
+ */
+export const STORE_NAMESPACE = 'core/wp-notifications';
+
+/**
+ * API_PATH WP Notification Feature rest api path.
+ */
+export const API_PATH = '/wp-notifications/v1/notifications/';
+
+/**
+ * The width of the notification hub.
+ */
+export const HUB_WIDTH = 320;
+
+/**
+ * The number of milliseconds in a week.
+ */
+export const WEEK_IN_MILLISECONDS = 604_800_000;
