--- src/scripts/constants.js
@@ -1,20 +1,0 @@
-/**
- * WP Notification Feature namespace.
- */
-export const STORE_NAMESPACE = /** @type {const} */ 'core/wp-notifications';
-
-/**
- * API_PATH WP Notification Feature rest api path.
- */
-export const API_PATH =
-	/** @type {const} */ '/wp-notifications/v1/notifications/';
-
-/**
- * The width of the notification hub.
- */
-export const HUB_WIDTH = /** @type {const} */ 320;
-
-/**
- * The number of milliseconds in a week.
- */
-export const WEEK_IN_MILLISECONDS = /** @type {const} */ ( 604800000 );
