--- src/wp-notifications.js
@@ -1,3 +1,0 @@
-import './scripts/';
-
-import './styles/wp-notifications.scss';
