+++ src/wp-notifications.scss
@@ -1,0 +1,20 @@
+/**
+ * WP Notification Feature Style
+ */
+
+/* Vars */
+@import "styles/vars";
+
+/* Components */
+@import "styles/components/notification";
+@import "styles/components/bullet";
+@import "components/unread-dot/style";
+
+// Dashboard notifications
+@import "styles/dashboard/notice";
+
+// Hub notifications
+@import "styles/hub/admin-bar";
+@import "styles/hub/layout";
+@import "styles/hub/elements";
+@import "styles/hub/notice";
