+++ webpack.config.js
@@ -1,5 +1,6 @@
-const defaultConfig = require( '@wordpress/scripts/config/webpack.config' );
 const path = require( 'path' );
+
+const defaultConfig = require( '@wordpress/scripts/config/webpack.config' );
 
 module.exports = {
 	...defaultConfig,
